grpc_chat/pyproject.toml
[tool.poetry]
name = "grpc-chat"
version = "0.1.0"
description = ""
authors = ["alexbuyan <sashabuyan@yandex.ru>"]
readme = "README.md"
packages = [{include = "grpc_chat", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
